--- packages/client/src/components/inscription/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./formInscription";
