+++ packages/client/src/pages/connexion/connexion.tsx
@@ -1,102 +1,333 @@
-import { useEffect, useState } from "react";
-import CardMedia from "@mui/material/CardMedia";
-import Typography from "@mui/material/Typography";
-import { useNavigate } from "react-router";
-import { Routes } from "../../app/routes";
-import { Categorie, ProduitWithPhoto } from "../../types/produits";
-
+import { Box, MenuItem, TextField, Typography } from "@mui/material";
+import { ValidationGroup } from "mui-validate";
+import React, {
+  ChangeEvent,
+  FormEvent,
+  FormEventHandler,
+  useContext,
+  useEffect,
+  useState,
+} from "react";
+import { useLocation } from "react-router";
+import axios from "../../axios";
+import { UserRoles } from "../../constants/roles";
+import UpLoad from "../../images/UpLoad.png";
+import img from "../../images/DevonRex.jpg";
+import AuthContext from "../../store/auth/AuthContextProvider";
+import { Roles } from "../../types/produits";
 import * as S from "./connexion.styled";
 
 type Props = {
-  element?: Categorie;
-  hidden?: boolean;
-  elProduit?: ProduitWithPhoto;
+  onSubmit: FormEventHandler<HTMLFormElement>;
 };
-const CartProduit: React.FC<Props> = ({
-  element,
-  hidden = false,
-  elProduit = undefined,
-}) => {
-  const navigate = useNavigate();
-  const [dataUrl, setDataUrl] = useState("");
 
-  useEffect(() => {
-    if (element?.photo_principale) {
-      setDataUrl(element?.photo_principale);
-    }
-    if (elProduit !== undefined && elProduit?.photo) {
-      setDataUrl(elProduit?.photo);
-    }
-  }, [elProduit, element]);
+const Connexion: React.FC<Props> = ({ onSubmit }) => {
+  const [validationNom, setValidationNom] = useState({
+    valid: false,
+    messages: [],
+    display: false,
+  });
+  const [validationPreNom, setValidationPreNom] = useState({
+    valid: false,
+    messages: [],
+    display: false,
+  });
+  const [validationEmail, setValidationEmail] = useState({
+    valid: false,
+    messages: [],
+    display: false,
+  });
 
-  const openCard = () => {
-    let categorieId;
-    if (typeof element === "number") categorieId = element;
-    else categorieId = element?.id ?? elProduit?.id_categorie;
-    const produitId = elProduit?.id ?? 0;
-    navigate(Routes.admin, {
-      state: { categorieId: categorieId, produitId: produitId },
-    });
+  const [passwordValue, setPasswordValue] = useState("");
+
+  const [validationPassword, setValidationPassword] = useState<{
+    valid: boolean;
+    messages: string[];
+    display: boolean;
+  }>({
+    valid: false,
+    messages: [],
+    display: false,
+  });
+
+  const location = useLocation();
+  const { authState } = useContext(AuthContext);
+  const [selectedRole, setSelectedRole] = useState(UserRoles.VISITEUR);
+  const [roles, setRoles] = useState<Roles[]>();
+  const [dataUrl, setDataUrl] = useState(UpLoad);
+
+  const currentPathArray = location.pathname.split("/");
+  const page = currentPathArray[currentPathArray.length - 1];
+  const userRole = authState.role;
+
+  const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
+    setSelectedRole(event.target.value as UserRoles);
   };
 
-  if (hidden) return null;
+  const validationForm =
+    validationNom.valid &&
+    validationPreNom.valid &&
+    validationEmail.valid &&
+    validationPassword.valid;
+
+  const addValidationForm = (event: FormEvent<HTMLFormElement>) => {
+    // Soumettre le formulaire si toutes les validations sont valides
+    if (
+      validationNom.valid &&
+      validationPreNom.valid &&
+      validationEmail.valid &&
+      validationPassword.valid
+    ) {
+      onSubmit(event);
+    }
+  };
+
+  useEffect(() => {
+    const fetchGetRoles = async () => {
+      try {
+        const response = await axios.get(`roles`);
+        setRoles(response.data.results[0]);
+      } catch (err) {
+        console.error(err);
+      }
+    };
+
+    fetchGetRoles();
+  }, []);
+
+  const handleUpload = async (event: ChangeEvent<HTMLInputElement>) => {
+    if (event.target.files && event.target.files.length > 0) {
+      setDataUrl(URL.createObjectURL(event.target.files[0]));
+    }
+  };
+
+  const isLengthValid = (password: string) => password.length >= 7;
+  const hasUpperCase = (password: string) => /[A-Z]/.test(password);
+  const hasLowerCase = (password: string) => /[a-z]/.test(password);
+  const hasSpecialChar = (password: string) =>
+    /[!@#$%^&*(),.?":{}|<>]/.test(password);
+
+  useEffect(() => {
+    const passwordValidations = [
+      {
+        condition: isLengthValid(passwordValue),
+        message: "au moins 7 caractères",
+      },
+      {
+        condition: hasUpperCase(passwordValue),
+        message: "au moins une majuscule",
+      },
+      {
+        condition: hasLowerCase(passwordValue),
+        message: "au moins une minuscule",
+      },
+      {
+        condition: hasSpecialChar(passwordValue),
+        message: "au moins un caractère spécial",
+      },
+    ];
+
+    const invalidMessages = passwordValidations
+      .filter((validation) => !validation.condition)
+      .map((validation) => validation.message);
+
+    setValidationPassword({
+      valid: invalidMessages.length === 0,
+      messages: invalidMessages,
+      display: true, // Afficher les messages d'erreur
+    });
+  }, [passwordValue]);
 
   return (
     <S.MainContainer>
-      <S.ImageButton
-        focusRipple
-        onClick={() => {
-          openCard();
-        }}
-      >
-        <CardMedia
-          component="img"
-          image={dataUrl}
-          alt={element?.categorie ?? elProduit?.titre}
-          sx={{
-            width: "99%",
-            height: "99%",
-            boxShadow: "14px 14px 14px red",
-          }}
-        />
-        <S.ImageBackdrop className="MuiImageBackdrop-root" />
-
-        <S.Image>
-          <S.FlexContainer>
-            <Typography
-              component="div"
-              variant="h2"
-              color="inherit"
-              sx={{
-                position: "relative",
-                p: 4,
-                pt: 2,
-                pb: (theme) => `calc(${theme.spacing(1)} + 6px)`,
-              }}
-            >
-              {elProduit?.categorie ?? element?.categorie ?? ""}
-              <S.ImageMarked className="MuiImageMarked-root" />
-            </Typography>
-            {elProduit?.titre ? (
+      <ValidationGroup>
+        <Box
+          component="form"
+          onSubmit={addValidationForm}
+          sx={{ height: "auto" }}
+        >
+          <input type="hidden" name="page" value={page} />
+          <S.FlexBox>
+            <S.ImgContainer>
+              <S.Img src={img} alt="" />
+              <S.StyledAvatar alt="user avatar" src={dataUrl} />
+              <S.StyledInput
+                type="file"
+                onChange={handleUpload}
+                name="avatar"
+              />
               <Typography
-                component="div"
                 variant="h2"
-                color="inherit"
                 sx={{
-                  position: "relative",
-                  p: 4,
-                  pt: 2,
-                  pb: (theme) => `calc(${theme.spacing(1)} + 6px)`,
+                  pt: { sm: 3, md: 6 },
+                  color: "primary.main",
+                  fontWeight: "900",
+                  gridRow: "2",
+                  gridColumn: "1",
                 }}
               >
-                {elProduit?.titre}
+                Choisissez un avatar
               </Typography>
-            ) : null}
-          </S.FlexContainer>
-        </S.Image>
-      </S.ImageButton>
+            </S.ImgContainer>
+            <S.StyledBox>
+              <S.FlexBoxTitle>
+                <Typography noWrap variant="body1" textAlign="center">
+                  Nom :
+                </Typography>
+                <Typography noWrap variant="body1" textAlign="center">
+                  Prénom :
+                </Typography>
+              </S.FlexBoxTitle>
+              <S.FlexBoxMain>
+                <S.Title>
+                  <Typography variant="body1" textAlign="center">
+                    Nom :
+                  </Typography>
+                </S.Title>
+                <TextField
+                  required
+                  type="text"
+                  placeholder="Entrez votre nom..."
+                  sx={{
+                    width: { xs: "60vw", md: "15vw" },
+                    backgroundColor: "secondary.main",
+                  }}
+                  name="nom"
+                  onChange={(event) =>
+                    setValidationNom({
+                      valid: !!event.target.value,
+                      messages: [],
+                      display: false,
+                    })
+                  }
+                />
+                <S.Title>
+                  <Typography noWrap variant="body1" textAlign="center">
+                    Prénom :
+                  </Typography>
+                </S.Title>
+                <TextField
+                  required
+                  type="text"
+                  placeholder="Entrez votre prénom.."
+                  sx={{
+                    width: { xs: "60vw", md: "15vw" },
+                    textAlign: "center",
+                    backgroundColor: "secondary.main",
+                  }}
+                  name="prenom"
+                  onChange={(event) =>
+                    setValidationPreNom({
+                      valid: !!event.target.value,
+                      messages: [],
+                      display: false,
+                    })
+                  }
+                />
+              </S.FlexBoxMain>
+              <div> &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</div>
+              <S.FlexBoxTitle>
+                <Typography variant="body1" textAlign="center">
+                  Role :
+                </Typography>
+                <Typography variant="body1" textAlign="center">
+                  Email :
+                </Typography>
+              </S.FlexBoxTitle>
+              <S.FlexBoxMain>
+                <S.Title>
+                  <Typography noWrap variant="body1" textAlign="center">
+                    Role :
+                  </Typography>
+                </S.Title>
+                <TextField
+                  id="outlined-select-role"
+                  select
+                  label="Select"
+                  name="role"
+                  value={selectedRole ? selectedRole : UserRoles.VISITEUR}
+                  defaultValue={UserRoles.VISITEUR}
+                  sx={{
+                    width: { xs: "60vw", md: "15vw" },
+                    textAlign: "center",
+                    backgroundColor: "secondary.main",
+                    mb: 2,
+                  }}
+                  disabled={userRole !== UserRoles.ADMINISTRATEUR}
+                  onChange={handleChange}
+                >
+                  {roles?.map((item, index) => (
+                    <MenuItem key={index} value={item.role}>
+                      {item.role}
+                    </MenuItem>
+                  ))}
+                </TextField>
+                <TextField
+                  required
+                  type="text"
+                  placeholder="Entrez votre email..."
+                  sx={{
+                    width: { xs: "60vw", md: "15vw" },
+                    textAlign: "center",
+                    backgroundColor: "secondary.main",
+                    mb: 2,
+                  }}
+                  name="email"
+                  onChange={(event) =>
+                    setValidationEmail({
+                      valid: /^[\w-.]+@([\w-]+\.)+[\w-]{2,4}$/.test(
+                        event.target.value
+                      ),
+                      messages: [],
+                      display: false,
+                    })
+                  }
+                />
+              </S.FlexBoxMain>
+              <Typography variant="body1" textAlign="center">
+                Mot de passe :
+              </Typography>
+              <TextField
+                required
+                type="password"
+                placeholder="Entrez votre mot de passe..."
+                sx={{
+                  width: { xs: "60vw", md: "36vw" },
+                  textAlign: "center",
+                  backgroundColor: "secondary.main",
+                }}
+                name="password"
+                value={passwordValue}
+                onChange={(event) => setPasswordValue(event.target.value)}
+                autoComplete="new-password"
+              />
+              {/* Affichage des messages de validation du mot de passe */}
+              <S.FlexContainer>
+                {validationPassword.display && (
+                  <Box>
+                    {validationPassword.messages.map((message, index) => (
+                      <Typography key={index} variant="body1" color="red">
+                        {message}
+                      </Typography>
+                    ))}
+                  </Box>
+                )}
+
+                <div> &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;</div>
+                <S.ButtonSubmit
+                  type="submit"
+                  disabled={!validationForm}
+                  color="primary"
+                >
+                  Submit
+                </S.ButtonSubmit>
+              </S.FlexContainer>
+            </S.StyledBox>
+          </S.FlexBox>
+        </Box>
+      </ValidationGroup>
     </S.MainContainer>
   );
 };
 
-export default CartProduit;
+export default Connexion;
